requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Computer Vision & Biometrics (Compatible versions)
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Note: For face recognition, you may need to install these separately:
# face-recognition (requires CMake and Visual Studio Build Tools on Windows)
# dlib (requires CMake and Visual Studio Build Tools on Windows)

# Alternative: Use simpler face detection with OpenCV instead
# mediapipe>=0.10.0  # Alternative for face detection

# Image Processing
# scikit-image>=0.22.0
# imutils>=0.5.4

# Cryptography
cryptography>=41.0.0
bcrypt>=4.1.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
aiofiles>=23.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Optional Database Drivers
# psycopg2-binary>=2.9.0  # For PostgreSQL
# pymysql>=1.1.0          # For MySQL

# Optional: Install these for full biometric functionality
# Requires Visual Studio Build Tools or pre-compiled wheels
# face-recognition>=1.3.0
# dlib>=19.24.0
